refactor: move GuantrCircuitBreakerError to errors module
Add full documentation for the error class and update the
createGuantr description to mention the thrown error.
