+++ src/index.ts
@@ -6,6 +6,7 @@ import type {
   GuantrResourceMap,
   GuantrOptions,
 } from './types';
+import { GuantrCircuitBreakerError } from './errors';
 import { InMemoryStorage } from './storage';
 import {
   getContextValue,
@@ -14,7 +15,11 @@ import {
   validateConditionForStrict,
 } from './utils';
 
-export { GuantrInvalidConditionError, GuantrInvalidConditionOperatorError } from './errors';
+export {
+  GuantrCircuitBreakerError,
+  GuantrInvalidConditionError,
+  GuantrInvalidConditionOperatorError,
+} from './errors';
 
 export type {
   GuantrMeta,
@@ -132,31 +137,6 @@ type CannotMethod<Meta extends GuantrMeta<GuantrResourceMap> | undefined> = {
     resource: ResourceKey,
   ): Promise<boolean>;
 };
-
-/**
- * Error thrown when the circuit breaker trips due to excessive rule iterations.
- * Indicates that the number of rules evaluated for a single permission check
- * exceeded the configured `maxRuleIterations` limit.
- */
-export class GuantrCircuitBreakerError extends Error {
-  /** The action being checked when the circuit breaker tripped. */
-  action: string;
-  /** The resource key being checked when the circuit breaker tripped. */
-  resource: string;
-  /** The configured iteration limit that was exceeded. */
-  limit: number;
-
-  constructor(action: string, resource: string, limit: number) {
-    super(
-      `[guantr] Circuit breaker tripped: rule iteration limit (${limit}) exceeded while evaluating action "${action}" on resource "${resource}". ` +
-        `Consider reducing the number of rules or increasing the \`maxRuleIterations\` option.`,
-    );
-    this.name = 'GuantrCircuitBreakerError';
-    this.action = action;
-    this.resource = resource;
-    this.limit = limit;
-  }
-}
 
 export class Guantr<
   Meta extends GuantrMeta<GuantrResourceMap> | undefined = undefined,
